More and harder enemies added, added must die boolean for enemies + speed is now float instead of int

+++ StarWars/Entity.cs
@@ -5,7 +5,7 @@ namespace StarWars
     class Entity:GameObject
     {
         //Speed for the entity when it moves
-        protected int speed;
+        protected float speed;
 
         //If the entity is alive or not
         protected bool alive = true;
@@ -16,7 +16,7 @@ namespace StarWars
         public bool Alive { get => alive; set => alive = value; }
         public int Hitpoints { get => hitpoints; set => hitpoints = value; }
 
-        public Entity(Texture2D texture, int hitboxX, int hitboxY, int speed):base(texture, hitboxX, hitboxY)
+        public Entity(Texture2D texture, int hitboxX, int hitboxY, float speed):base(texture, hitboxX, hitboxY)
         {
             //Movement speed of the entity
             this.speed = speed;
